Frittie/requirements.txt
Django==1.4
Markdown==2.1.1
MySQL-python==1.2.3
PIL==1.1.7
South==0.7.5
Twisted==12.1.0
Whoosh==2.4.1
amqplib==1.0.2
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.12
celery==3.0.5
demjson==1.6
distribute==0.6.24
django-allauth==0.6.0
django-celery==3.0.4
django-dajax==0.8.4
django-dajaxice==0.2
django-email-confirmation==0.2
django-facebook==4.0.9
django-haystack==1.2.7
django-simple-friends==0.4
django-tastypie==0.9.11
django-uni-form==0.9.0
django-user-streams==0.5.0
httplib2==0.7.4
kombu==2.3.2
mimeparse==0.1.3
morbid==0.8.7.3
oauth2==1.5.211
orbited==0.7.10
pgmagick==0.5.1
python-dateutil==1.5
python-openid==2.2.5
redis==2.4.13
sorl-thumbnail==11.12
stomp.py==3.1.3
stomper==0.2.4
wadofstuff-django-serializers==1.1.0
wsgiref==0.1.2
yolk==0.4.3
zope.interface==4.0.1
